Add files via upload
So this is the one that messed up.  I added in the instructions, to pop up in a separate window before the game starts.  It seemed to work at first, but then as I kept working with it all the other buttons stopped appearing.  I'm just getting a white box, like there's nothing in it.  So while I've been trying to figure this out, I made a separate file that I've been using to work on the canvas so that it doesn't interfere with anything